Verification de la bonne execution de l'application

+++ voteapp-webapp/src/main/java/fr/sgr/formation/voteapp/elections/services/ElectionsServices.java
@@ -11,6 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import fr.sgr.formation.voteapp.elections.modele.Election;
 import fr.sgr.formation.voteapp.elections.services.ElectionInvalideException.ErreurElection;
 import fr.sgr.formation.voteapp.notifications.services.NotificationsServices;
+import fr.sgr.formation.voteapp.utilisateurs.services.AuthentificationUtilisateursServices;
 import lombok.extern.slf4j.Slf4j;
 
 @Service
@@ -27,6 +28,11 @@ public class ElectionsServices {
 	@Autowired
 	private EntityManager entityManager;
 
+	@Autowired
+	private AuthentificationUtilisateursServices authentificationUtilisateursServices;
+	@Autowired
+	private VerificationCreateurService verificationCreateurService;
+
 	/**
 	 * Crée une nouvelle élection sur le système.
 	 * 
@@ -37,30 +43,37 @@ public class ElectionsServices {
 	 *             Levée si l'election est invalide.
 	 */
 	@Transactional(propagation = Propagation.REQUIRED)
-	public Election creer(Election election) throws ElectionInvalideException {
-		log.info("=====> Création de l'election : {}.", election);
+	public Election creer(Election election, String idUser) throws ElectionInvalideException {
+		boolean gerant = authentificationUtilisateursServices.gerantVerif(idUser);
+		if (gerant) {
 
-		if (election == null) {
-			throw new ElectionInvalideException(ErreurElection.ELECTION_OBLIGATOIRE);
+			log.info("=====> Création de l'election : {}.", election);
+
+			if (election == null) {
+				throw new ElectionInvalideException(ErreurElection.ELECTION_OBLIGATOIRE);
+			}
+
+			/** Validation de l'existance de l'election. */
+			if (rechercherParTitre(election.getTitre()) != null) {
+				throw new ElectionInvalideException(ErreurElection.ELECTION_EXISTANTE);
+			}
+
+			/**
+			 * Validation de l'election: lève une exception si l'election est
+			 * invalide.
+			 */
+			validationServices.validerElection(election);
+
+			/** Notification de l'événement de création */
+			notificationsServices.notifier("Création de l'election: " + election.toString());
+
+			/** Persistance de l'election. */
+			entityManager.persist(election);
+		} else {
+			notificationsServices
+					.notifier("Impossible de créer l'élection " + election.getTitre()
+							+ " car vous n'êtes pas gérant");
 		}
-
-		/** Validation de l'existance de l'election. */
-		if (rechercherParTitre(election.getTitre()) != null) {
-			throw new ElectionInvalideException(ErreurElection.ELECTION_EXISTANTE);
-		}
-
-		/**
-		 * Validation de l'election: lève une exception si l'election est
-		 * invalide.
-		 */
-		validationServices.validerElection(election);
-
-		/** Notification de l'événement de création */
-		notificationsServices.notifier("Création de l'election: " + election.toString());
-
-		/** Persistance de l'election. */
-		entityManager.persist(election);
-
 		return election;
 	}
 
@@ -74,28 +87,48 @@ public class ElectionsServices {
 	 *             Levée si l'election est invalide.
 	 */
 	@Transactional(propagation = Propagation.REQUIRED)
-	public Election cloturer(Election election) throws ElectionInvalideException {
-		log.info("=====> cloture de l'election : {}.", election);
+	public Election cloturer(String titre, String idUser) throws ElectionInvalideException {
+		/** Validation de l'existance de l'election. */
+		if (rechercherParTitre(titre) == null) {
+			throw new ElectionInvalideException(ErreurElection.ELECTION_NON_EXISTANTE);
+		}
+
+		Election election = entityManager.find(Election.class, titre);
 
 		if (election == null) {
 			throw new ElectionInvalideException(ErreurElection.ELECTION_OBLIGATOIRE);
 		}
 
-		/** Validation de l'existance de l'election. */
-		if (rechercherParTitre(election.getTitre()) == null) {
-			throw new ElectionInvalideException(ErreurElection.ELECTION_NON_EXISTANTE);
+		log.info("=====> Cloture de l'election : {}.", election);
+
+		boolean gerant = authentificationUtilisateursServices.gerantVerif(idUser);
+		boolean createur = verificationCreateurService.createurVerif(idUser, titre);
+
+		if (gerant) {
+			// Supprime l'utilisateur de la base si il existe
+			if (createur) {
+				/**
+				 * Validation de l'election: lève une exception si l'election
+				 * est invalide.
+				 */
+				validationServices.validerElection(election);
+
+				/** Notification de l'événement de cloture */
+				notificationsServices.notifier("Cloture de l'election: " + election.toString());
+				entityManager.remove(election);
+				election.setActiveElection(false);
+				creer(election, idUser);
+
+			} else {
+				notificationsServices
+						.notifier("Impossible de supprimer l'élection " + titre + " car vous n'êtes pas le créateur");
+			}
+
+		} else {
+			notificationsServices
+					.notifier("Impossible de supprimer l'élection " + titre + " car vous n'êtes pas gérant");
+
 		}
-
-		election.setActiveElection(false);
-
-		/**
-		 * Validation de l'election: lève une exception si l'election est
-		 * invalide.
-		 */
-		validationServices.validerElection(election);
-
-		/** Notification de l'événement de création */
-		notificationsServices.notifier("Cloture de l'election: " + election.toString());
 
 		/** Persistance de l'election. */
 		// entityManager.persist(election);
@@ -104,7 +137,7 @@ public class ElectionsServices {
 	}
 
 	/**
-	 * Cloture uen élection
+	 * Modifier uen élection
 	 * 
 	 * @param election
 	 *            Election à cloturer.
@@ -113,7 +146,67 @@ public class ElectionsServices {
 	 *             Levée si l'election est invalide.
 	 */
 	@Transactional(propagation = Propagation.REQUIRED)
-	public Election supprimer(String titre) throws ElectionInvalideException {
+	public Election modifier(String titre, String idUser) throws ElectionInvalideException {
+		/** Validation de l'existance de l'election. */
+		if (rechercherParTitre(titre) == null) {
+			throw new ElectionInvalideException(ErreurElection.ELECTION_NON_EXISTANTE);
+		}
+
+		Election election = entityManager.find(Election.class, titre);
+
+		if (election == null) {
+			throw new ElectionInvalideException(ErreurElection.ELECTION_OBLIGATOIRE);
+		}
+
+		log.info("=====> Cloture de l'election : {}.", election);
+
+		boolean gerant = authentificationUtilisateursServices.gerantVerif(idUser);
+		boolean createur = verificationCreateurService.createurVerif(idUser, titre);
+
+		if (gerant) {
+			// Supprime l'utilisateur de la base si il existe
+			if (createur) {
+				/**
+				 * Validation de l'election: lève une exception si l'election
+				 * est invalide.
+				 */
+				validationServices.validerElection(election);
+
+				/** Notification de l'événement de cloture */
+				notificationsServices.notifier("Modification de l'election: " + election.toString());
+				entityManager.remove(election);
+				election.setActiveElection(false);
+
+				creer(election, idUser);
+
+			} else {
+				notificationsServices
+						.notifier("Impossible de supprimer l'élection " + titre + " car vous n'êtes pas le créateur");
+			}
+
+		} else {
+			notificationsServices
+					.notifier("Impossible de supprimer l'élection " + titre + " car vous n'êtes pas gérant");
+
+		}
+
+		/** Persistance de l'election. */
+		// entityManager.persist(election);
+
+		return election;
+	}
+
+	/**
+	 * Supprimer uen élection
+	 * 
+	 * @param election
+	 *            Election à cloturer.
+	 * @return Election cloturée.
+	 * @throws ElectionInvalideException
+	 *             Levée si l'election est invalide.
+	 */
+	@Transactional(propagation = Propagation.REQUIRED)
+	public Election supprimer(String titre, String idUser) throws ElectionInvalideException {
 		/** Validation de l'existance de l'election. */
 		if (rechercherParTitre(titre) == null) {
 			throw new ElectionInvalideException(ErreurElection.ELECTION_NON_EXISTANTE);
@@ -125,24 +218,36 @@ public class ElectionsServices {
 			throw new ElectionInvalideException(ErreurElection.ELECTION_OBLIGATOIRE);
 		}
 
-		// Supprime l'utilisateur de la base si il existe
-		if (election != null) {
-			entityManager.remove(election);
-			/** Notification de l'événement de création */
-			notificationsServices.notifier("Suppression de l'élection: " + election.toString());
+		boolean gerant = authentificationUtilisateursServices.gerantVerif(idUser);
+		boolean createur = verificationCreateurService.createurVerif(idUser, titre);
+
+		if (gerant) {
+			// Supprime l'utilisateur de la base si il existe
+			if (createur) {
+				if (election != null) {
+					/**
+					 * Validation de l'election: lève une exception si
+					 * l'election est invalide.
+					 */
+					validationServices.validerElection(election);
+
+					/** Notification de l'événement de création */
+					notificationsServices.notifier("Cloture de l'election: " + election.toString());
+					entityManager.remove(election);
+					/** Notification de l'événement de création */
+					notificationsServices.notifier("Suppression de l'élection: " + election.toString());
+
+				}
+			} else {
+				notificationsServices
+						.notifier("Impossible de supprimer l'élection " + titre + " car vous n'êtes pas le créateur");
+			}
+
 		} else {
-			/** Notification de l'événement de création */
-			notificationsServices.notifier("Impossible de supprimer l'élection " + titre + " car il n'existe pas.");
+			notificationsServices
+					.notifier("Impossible de supprimer l'élection " + titre + " car vous n'êtes pas gérant");
+
 		}
-
-		/**
-		 * Validation de l'election: lève une exception si l'election est
-		 * invalide.
-		 */
-		validationServices.validerElection(election);
-
-		/** Notification de l'événement de création */
-		notificationsServices.notifier("Cloture de l'election: " + election.toString());
 
 		/** Persistance de l'election. */
 		// entityManager.persist(election);
